requirements.txt
pandas
scikit-learn
streamlit
plotly_express
pycountry
joblib
requests 